# control/docs/swagger.json
{
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "The self-service API: sign in, manage your profile and your personal access tokens, and create and run VMs.\n\nTwo credentials reach these routes. A browser sends the httpOnly `access_token` cookie. A script sends a personal access token as `Authorization: Bearer dpat_...` — mint one under Settings, or with `POST /me/tokens` from a signed-in session.\n\nA personal access token does everything you can do with your own VMs. It can never reach `/api/v1/admin/*`, even when it belongs to an admin, and it cannot manage tokens: those routes take a password-backed session, so a leaked token cannot mint its successor.\n\nAdmin and client-enrollment routes are deliberately absent here. The first is browser-only by design; the second is a protocol between the control plane and its hosts rather than an API.",
        "title": "dummie control API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/integrations": {
            "get": {
                "responses": {},
                "summary": "Integrations you own",
                "tags": [
                    "integrations"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {},
                "summary": "Create an integration",
                "tags": [
                    "integrations"
                ]
            }
        },
...
        "/integrations/{id}": {
            "delete": {
                "responses": null,
                "summary": "Delete an integration",
                "tags": [
                    "integrations"
                ]
            },
            "get": {
                "responses": {},
                "summary": "One integration",
                "tags": [
                    "integrations"
                ]
            },
            "patch": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {},
                "summary": "Update an integration's flags",
                "tags": [
                    "integrations"
                ]
            }
        },
...
        "/integrations/{id}/repos": {
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": null,
                "summary": "Set which repositories an integration covers",
                "tags": [
                    "integrations"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}